+++ django_peeringdb/__init__.py
@@ -1,3 +1,0 @@
-import pkg_resources
-
-__version__ = pkg_resources.require("django_peeringdb")[0].version